refactor checker to new project struct
