--- checker/test/.go-arch-lint-v1.yml
@@ -1,23 +1,0 @@
-version: 1
-allow:
-  depOnAnyVendor: true
-excludeFiles:
-  - "^.*_test\\.go$"
-components:
-  a:
-    in: a
-  b:
-    in: b/*/target
-  c:
-    in: c/**
-  utils:
-    in: utils/**
-commonComponents:
-  - utils
-deps:
-  a:
-    mayDependOn:
-      - b
-  b:
-    mayDependOn:
-      - c
