+++ internal/app/internal/container/cmd_check.go
@@ -15,7 +15,7 @@ func (c *Container) ProvideCheckCommand() *cobra.Command {
 func (c *Container) provideCheckCommandAssembler() *check.CommandAssembler {
 	return check.NewCheckCommandAssembler(func(input models.FlagsCheck) error {
 		return c.ProvideRenderer().RenderModel(
-			c.provideCheckService(input).Behave(),
+			c.provideCheckService(input).Behave(input.MaxWarnings),
 		)
 	})
 }
