+++ internal/app/internal/container/services.go
@@ -21,6 +21,7 @@ func (c *Container) provideCheckService(input models.FlagsCheck) *check.Service 
 			input.ModuleName,
 			input.GoArchFilePath,
 		),
+		c.provideSpecChecker(),
 		c.provideReferenceRender(),
 	)
 }
