+++ internal/models/resolved_file.go
