+++ internal/services/check/service.go
@@ -9,34 +9,82 @@ import (
 
 type Service struct {
 	specAssembler   SpecAssembler
+	specChecker     SpecChecker
 	referenceRender ReferenceRender
 }
 
 func NewService(
 	specAssembler SpecAssembler,
+	specChecker SpecChecker,
 	referenceRender ReferenceRender,
 ) *Service {
 	return &Service{
 		specAssembler:   specAssembler,
+		specChecker:     specChecker,
 		referenceRender: referenceRender,
 	}
 }
 
-func (s *Service) Behave() (models.Check, error) {
+func (s *Service) Behave(maxWarnings int) (models.Check, error) {
 	spec, err := s.specAssembler.Assemble()
 	if err != nil {
 		return models.Check{}, fmt.Errorf("failed to assemble spec: %w", err)
 	}
 
+	result := s.specChecker.Check(spec)
+	result = s.limitResults(result, maxWarnings)
+
 	model := models.Check{
 		ModuleName:             spec.ModuleName.Value(),
 		DocumentNotices:        s.assembleNotice(spec.Integrity),
-		ArchHasWarnings:        false,
-		ArchWarningsDependency: []models.CheckArchWarningDependency{},
-		ArchWarningsMatch:      []models.CheckArchWarningMatch{},
+		ArchHasWarnings:        s.resultsHasWarnings(result),
+		ArchWarningsDependency: result.DependencyWarnings,
+		ArchWarningsMatch:      result.MatchWarnings,
 	}
 
 	return model, nil
+}
+
+func (s *Service) limitResults(result models.CheckResult, maxWarnings int) models.CheckResult {
+	totalCount := 0
+	limitedResults := models.CheckResult{
+		DependencyWarnings: []models.CheckArchWarningDependency{},
+		MatchWarnings:      []models.CheckArchWarningMatch{},
+	}
+
+	// append deps
+	for _, notice := range result.DependencyWarnings {
+		if totalCount >= maxWarnings {
+			break
+		}
+
+		limitedResults.DependencyWarnings = append(limitedResults.DependencyWarnings, notice)
+		totalCount++
+	}
+
+	// append not matched
+	for _, notice := range result.MatchWarnings {
+		if totalCount >= maxWarnings {
+			break
+		}
+
+		limitedResults.MatchWarnings = append(limitedResults.MatchWarnings, notice)
+		totalCount++
+	}
+
+	return limitedResults
+}
+
+func (s *Service) resultsHasWarnings(result models.CheckResult) bool {
+	if len(result.DependencyWarnings) > 0 {
+		return true
+	}
+
+	if len(result.MatchWarnings) > 0 {
+		return true
+	}
+
+	return false
 }
 
 func (s *Service) assembleNotice(integrity speca.Integrity) []models.CheckNotice {
